+++ decidim-proposals/app/commands/decidim/proposals/vote_proposal.rb
@@ -20,6 +20,8 @@ module Decidim
       #
       # Returns nothing.
       def call
+        return broadcast(:invalid) if @proposal.maximum_votes_reached?
+
         build_proposal_vote
         return broadcast(:invalid) unless vote.valid?