+++ decidim-proposals/app/models/decidim/proposals/proposal.rb
@@ -113,6 +113,25 @@ module Decidim
       def official?
         author.nil?
       end
+
+      # Public: The maximum amount of votes allowed for this proposal.
+      #
+      # Returns an Integer with the maximum amount of votes, nil otherwise.
+      def maximum_votes
+        maximum_votes = feature.settings.maximum_votes_per_proposal
+        return nil if maximum_votes.zero?
+
+        maximum_votes
+      end
+
+      # Public: The maximum amount of votes allowed for this proposal. 0 means infinite.
+      #
+      # Returns true if reached, false otherwise.
+      def maximum_votes_reached?
+        return false unless maximum_votes
+
+        votes.count >= maximum_votes
+      end
     end
   end
 end