+++ decidim-proposals/lib/decidim/proposals/test/factories.rb
@@ -68,6 +68,14 @@ FactoryGirl.define do
         }
       end
     end
+
+    trait :with_maximum_votes_per_proposal do
+      settings do
+        {
+          maximum_votes_per_proposal: 1
+        }
+      end
+    end
   end
 
   factory :proposal, class: Decidim::Proposals::Proposal do