+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - **decidim-core**: Add fields to customize the Call To Action homepage button text and URL [\#2053](https://github.com/decidim/decidim/pull/2053)
 - **decidim-meetings**: Participatory admins can invite users to join a meeting. [\#1879](https://github.com/decidim/decidim/pull/1879)
 - **decidim-proposals**: Show votes in the proposals table in admin when votes are enabled [\#2011](https://github.com/decidim/decidim/pull/2011)
+- **decidim-proposals**: Allow setting the maximum amount of votes a proposal can receive [\#2054](https://github.com/decidim/decidim/pull/2054).
 
 **Changed**